feat: edit form to High Cohesion Quality dquote> แยก src/components , hook , constant , app.jsx dquote> แบ่งส่วนการทำงานใน components , ActionButtons , FormField , MovieSelector , SummaryPage , UserForm
